+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # CHANGELOG
 
-## Unreleased changes
+## 0.0.8
+* サーバー起動時にホスト名を指定しないとホストPC以外でアクセスできなくなる問題の修正
+* 内部APIのエンドポイントを以下のように変更
+  * /config/engine => /config/content.json
+  * /config/engine/raw => /config/content.raw.json
 * サーバーのコンソール上にログ出力するオプションを追加
   * オプション未使用時は、warn と error 以外のログ出力が行われないようになった